integración de las 3 apis

+++ Criollos/Pedidos/src/main/java/com/pedidos/Pedidos/infrastructure/driver_adapters/jpa_repository/DetallePedidoData.java
@@ -30,7 +30,7 @@ public class DetallePedidoData {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Long productoId;
+    private Integer productoId;
 
     private String nombreProducto;
